requirements.txt
aniso8601==8.1.0
attrs==20.3.0
click==7.1.2
Flask==1.1.2
flask-restx==0.2.0
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
pyrsistent==0.17.3
python-dotenv==0.15.0
pytz==2020.5
PyYAML==5.3.1
six==1.15.0
SQLAlchemy==1.3.22
SQLAlchemy-Utils==0.36.8
Werkzeug==1.0.1
wgconfig==0.2.2
